+++ python/GafferSceneTest/__init__.py
@@ -173,6 +173,9 @@ from .MeshSplitTest import MeshSplitTest
 from .FramingConstraintTest import FramingConstraintTest
 from .MeshNormalsTest import MeshNormalsTest
 from .ImageScatterTest import ImageScatterTest
+from .RenderPassesTest import RenderPassesTest
+from .DeleteRenderPassesTest import DeleteRenderPassesTest
+from .RenderPassWedgeTest import RenderPassWedgeTest
 
 from .IECoreScenePreviewTest import *
 from .IECoreGLPreviewTest import *